fix: 채팅방 목록 레이아웃 문제 해결
#94
